refactor: move state to state.ts
