+++ src/app/features/forms/common/utils/isUserInputMarginError/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState as LPFormSliceState } from '../../../lp-form/reducer';
+import { SliceState as LPFormSliceState } from '../../../lp-form/state';
 import { SliceState as SwapFormSliceState } from '../../../swap-form/reducer';
 
 export const isUserInputMarginError = (
