+++ src/app/features/forms/lp-form/utils/getAvailableMargin/index.ts
@@ -1,7 +1,7 @@
 import { Draft } from '@reduxjs/toolkit';
 import { Position } from '@voltz-protocol/v1-sdk';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 import { getProspectiveLpNotional } from '../getProspectiveLpNotional';
 import { hasExistingPosition } from '../hasExistingPosition';
