+++ src/app/features/forms/lp-form/utils/getDefaultLpFixedHigh/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const getDefaultLpFixedHigh = (state: Draft<SliceState>): number => {
   // todo: layer in smarter dynamic default range logic in here
