+++ src/app/features/forms/lp-form/utils/getDefaultLpFixedLow/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const getDefaultLpFixedLow = (state: Draft<SliceState>): number => {
   // todo: layer in smarter dynamic default range logic in here
