+++ src/app/features/forms/lp-form/utils/getEditPositionNotional/index.test.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 import { getEditPositionNotional } from './index';
 
 describe('getEditPositionNotional', () => {
