+++ src/app/features/forms/lp-form/utils/getProspectiveLpFixedLow/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 import { isUserInputFixedRangeError } from '../isUserInputFixedRangeError';
 
 export const getProspectiveLpFixedLow = (state: Draft<SliceState>): number | null => {
