+++ src/app/features/forms/lp-form/utils/getProspectiveLpMargin/index.ts
@@ -1,7 +1,7 @@
 import { Draft } from '@reduxjs/toolkit';
 
 import { isUserInputMarginError } from '../../../common/utils';
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const getProspectiveLpMargin = (state: Draft<SliceState>): number => {
   if (isUserInputMarginError(state)) {
