+++ src/app/features/forms/lp-form/utils/getSelectedPositionNotional/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const getSelectedPositionNotional = (state: Draft<SliceState>) => {
   return state.selectedPosition?.notional ?? 0;
