+++ src/app/features/forms/lp-form/utils/hasExistingPosition/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const hasExistingPosition = (state: Draft<SliceState>): boolean => {
   return state.selectedPosition !== null;
