+++ src/app/features/forms/lp-form/utils/isUserInputFixedRangeError/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const isUserInputFixedRangeError = (state: Draft<SliceState>): boolean => {
   return state.userInput.fixedRange.error !== null;
