+++ src/app/features/forms/lp-form/utils/updateLeverageOptionsAfterGetPoolLpInfo/index.ts
@@ -2,7 +2,7 @@ import { Draft } from '@reduxjs/toolkit';
 
 import { formatNumber } from '../../../../../../utilities/number';
 import { calculateLeverageOptions } from '../../../common/utils';
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const updateLeverageOptionsAfterGetPoolLpInfo = (state: Draft<SliceState>): void => {
   const maxLeverage = formatNumber(Math.floor(state.poolLpInfo.maxLeverage), 0, 0);
